docker: place container, pipeline and log message into different files
