+++ pkg/docker/driver.go
@@ -1,21 +1,14 @@
 package docker
 
 import (
-	"bytes"
 	"context"
-	"encoding/binary"
-	"encoding/json"
 	"fmt"
-	"io"
-	"os"
 	"path"
 	"strings"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/robfig/cron"
-	"github.com/tonistiigi/fifo"
 
 	"golang.org/x/sync/errgroup"
 
@@ -24,10 +17,7 @@ import (
 	"github.com/docker/docker/daemon/logger"
 
 	"github.com/rchicoli/docker-log-elasticsearch/pkg/elasticsearch"
-	"github.com/rchicoli/docker-log-elasticsearch/pkg/extension/grok"
 	"github.com/rchicoli/docker-log-elasticsearch/pkg/regex"
-
-	protoio "github.com/gogo/protobuf/io"
 )
 
 const (
@@ -37,89 +27,8 @@ const (
 // Driver ...
 type Driver struct {
 	mu   *sync.Mutex
-	logs map[string]*container
 	ctx  context.Context
-}
-
-type container struct {
-	cron      *cron.Cron
-	esClient  elasticsearch.Client
-	indexName string
-	info      logger.Info
-	logger    *log.Entry
-	pipeline  pipeline
-	stream    io.ReadCloser
-}
-
-type pipeline struct {
-	group    *errgroup.Group
-	inputCh  chan logdriver.LogEntry
-	outputCh chan LogMessage
-}
-
-// LogMessage ...
-type LogMessage struct {
-	logdriver.LogEntry
-	logger.Info
-
-	GrokLine map[string]string
-}
-
-// MarshalJSON ...
-func (l LogMessage) MarshalJSON() ([]byte, error) {
-	return json.Marshal(
-		struct {
-
-			// docker/daemon/logger/Info
-			Config              map[string]string `json:"config,omitempty"`
-			ContainerID         string            `json:"containerID,omitempty"`
-			ContainerName       string            `json:"containerName,omitempty"`
-			ContainerEntrypoint string            `json:"containerEntrypoint,omitempty"`
-			ContainerArgs       []string          `json:"containerArgs,omitempty"`
-			ContainerImageID    string            `json:"containerImageID,omitempty"`
-			ContainerImageName  string            `json:"containerImageName,omitempty"`
-			ContainerCreated    *time.Time        `json:"containerCreated,omitempty"`
-			ContainerEnv        []string          `json:"containerEnv,omitempty"`
-			ContainerLabels     map[string]string `json:"containerLabels,omitempty"`
-			LogPath             string            `json:"logPath,omitempty"`
-			DaemonName          string            `json:"daemonName,omitempty"`
-
-			//  api/types/plugin/logdriver/LogEntry
-			Line     string    `json:"message,omitempty"` // []byte to string
-			Source   string    `json:"source"`
-			TimeNano time.Time `json:"timestamp"` // int64 to Time
-			Partial  bool      `json:"partial"`
-
-			GrokLine map[string]string `json:"grok,omitempty"`
-		}{
-			Config:              l.Config,
-			ContainerID:         l.ContainerID,
-			ContainerName:       l.ContainerName,
-			ContainerEntrypoint: l.ContainerEntrypoint,
-			ContainerArgs:       l.ContainerArgs,
-			ContainerImageID:    l.ContainerImageID,
-			ContainerImageName:  l.ContainerImageName,
-			ContainerCreated:    l.timeOmityEmpty(),
-			ContainerEnv:        l.ContainerEnv,
-			ContainerLabels:     l.ContainerLabels,
-			LogPath:             l.LogPath,
-			DaemonName:          l.DaemonName,
-
-			GrokLine: l.GrokLine,
-
-			Line:     string(l.Line),
-			Source:   l.Source,
-			TimeNano: time.Unix(0, l.TimeNano),
-			Partial:  l.Partial,
-		})
-
-}
-
-func (l LogMessage) timeOmityEmpty() *time.Time {
-	if l.ContainerCreated.IsZero() {
-		return nil
-	}
-	return &l.ContainerCreated
+	logs map[string]*container
 }
 
 // NewDriver returns a pointer to driver
@@ -130,46 +39,9 @@ func NewDriver() *Driver {
 	}
 }
 
-// newContainer stores the container's configuration in memory
-// and returns a pointer to the container
-func (d *Driver) newContainer(ctx context.Context, file string) (*container, error) {
-
-	filename := path.Base(file)
-	log.WithField("fifo", file).Debug("created fifo file")
-
-	d.mu.Lock()
-	if _, exists := d.logs[filename]; exists {
-		return nil, fmt.Errorf("error: a logger for this container already exists: %s", filename)
-	}
-	d.mu.Unlock()
-
-	f, err := fifo.OpenFifo(ctx, file, syscall.O_RDONLY, 0700)
-	if err != nil {
-		return nil, fmt.Errorf("could not open fifo: %q", err)
-	}
-
-	d.mu.Lock()
-	c := &container{stream: f}
-	d.logs[filename] = c
-	d.mu.Unlock()
-
-	return c, nil
-}
-
-// getContainer retrieves the container's configuration from memory
-func (d *Driver) getContainer(file string) (*container, error) {
-
-	filename := path.Base(file)
-
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	c, exists := d.logs[filename]
-	if !exists {
-		return nil, fmt.Errorf("error: logger not found for socket ID: %v", file)
-	}
-
-	return c, nil
+// Name return the docker plugin name
+func (d *Driver) Name() string {
+	return name
 }
 
 // StartLogging implements the docker plugin interface
@@ -226,192 +98,6 @@ func (d *Driver) StartLogging(file string, info logger.Info) error {
 	return nil
 }
 
-// Read reads messages from proto buffer
-func (d *Driver) Read(ctx context.Context, file string) error {
-
-	c, err := d.getContainer(file)
-	if err != nil {
-		return err
-	}
-
-	c.pipeline.group.Go(func() error {
-
-		dec := protoio.NewUint32DelimitedReader(c.stream, binary.BigEndian, 1e6)
-		defer func() {
-			c.logger.Info("closing docker stream")
-			dec.Close()
-			close(c.pipeline.inputCh)
-		}()
-
-		var buf logdriver.LogEntry
-		var err error
-
-		for {
-			if err = dec.ReadMsg(&buf); err != nil {
-				if err == io.EOF {
-					c.logger.Debug("shutting down reader eof")
-					return nil
-				}
-				// the connection has been closed
-				// stop looping and close the input channel
-				// read /proc/self/fd/6: file already closed
-				if strings.Contains(err.Error(), os.ErrClosed.Error()) {
-					c.logger.WithError(err).Debug("shutting down fifo: closed by the writer")
-					break
-				}
-				if err != nil {
-					// the connection has been closed
-					// stop looping and closing the input channel
-					// read /proc/self/fd/6: file already closed
-					c.logger.WithError(err).Debug("shutting down fifo")
-					break
-					// do not return, otherwise group.Go closes the pipeline
-					// return err
-				}
-
-				dec = protoio.NewUint32DelimitedReader(c.stream, binary.BigEndian, 1e6)
-			}
-
-			// in case docker run command throws lots empty line messages
-			if len(bytes.TrimSpace(buf.Line)) == 0 {
-				c.logger.WithField("line", string(buf.Line)).Debug("trim space")
-				continue
-			}
-
-			select {
-			case c.pipeline.inputCh <- buf:
-			case <-ctx.Done():
-				c.logger.WithError(ctx.Err()).Error("closing read pipeline")
-				return ctx.Err()
-			}
-			buf.Reset()
-		}
-
-		return nil
-	})
-
-	return nil
-}
-
-// Stats shows metrics related to the bulk service
-// func (d *Driver) Stats(filename string, config Configuration) error {
-// TODO: create metrics from stats
-// d.pipeline.group.Go(func() error {
-// 	stats := d.esClient.Stats()
-
-// 	fields := log.Fields{
-// 		"flushed":   stats.Flushed,
-// 		"committed": stats.Committed,
-// 		"indexed":   stats.Indexed,
-// 		"created":   stats.Created,
-// 		"updated":   stats.Updated,
-// 		"succeeded": stats.Succeeded,
-// 		"failed":    stats.Failed,
-// 	}
-
-// 	for i, w := range stats.Workers {
-// 		fmt.Printf("Worker %d: Number of requests queued: %d\n", i, w.Queued)
-// 		fmt.Printf("           Last response time       : %v\n", w.LastDuration)
-// 		fields[fmt.Sprintf("w%d.queued", i)] = w.Queued
-// 		fields[fmt.Sprintf("w%d.lastduration", i)] = w.LastDuration
-// 	}
-// })
-// }
-
-// Parse filters line messages
-func (d *Driver) Parse(ctx context.Context, file, fields, grokMatch, grokPattern, grokPatternFrom, grokPatternSplitter string, grokNamedCapture bool) error {
-
-	c, err := d.getContainer(file)
-	if err != nil {
-		return err
-	}
-
-	c.pipeline.group.Go(func() error {
-		defer close(c.pipeline.outputCh)
-
-		groker, err := grok.NewGrok(grokMatch, grokPattern, grokPatternFrom, grokPatternSplitter, grokNamedCapture)
-		if err != nil {
-			return err
-		}
-
-		var logMessage string
-		// custom log message fields
-		msg := getLogMessageFields(fields, c.info)
-
-		for m := range c.pipeline.inputCh {
-
-			logMessage = string(m.Line)
-
-			// create message
-			msg.Source = m.Source
-			msg.Partial = m.Partial
-			msg.TimeNano = m.TimeNano
-
-			// TODO: create a PR to grok upstream for parsing bytes
-			// so that we avoid having to convert the message to string
-			msg.GrokLine, msg.Line, err = groker.ParseLine(grokMatch, logMessage, m.Line)
-			if err != nil {
-				c.logger.WithError(err).Error("could not parse line with grok")
-			}
-
-			select {
-			case c.pipeline.outputCh <- msg:
-			case <-ctx.Done():
-				c.logger.WithError(ctx.Err()).Error("closing parse pipeline")
-				return ctx.Err()
-			}
-
-		}
-
-		return nil
-	})
-
-	return nil
-}
-
-// Log sends messages to Elasticsearch Bulk Service
-func (d *Driver) Log(ctx context.Context, file string, config Configuration) error {
-
-	c, err := d.getContainer(file)
-	if err != nil {
-		return err
-	}
-
-	c.pipeline.group.Go(func() error {
-
-		err := c.esClient.NewBulkProcessorService(ctx, config.Bulk.workers, config.Bulk.actions, config.Bulk.size, config.Bulk.flushInterval, config.Bulk.stats)
-		if err != nil {
-			c.logger.WithError(err).Error("could not create bulk processor")
-		}
-
-		defer func() {
-			if err := c.esClient.Flush(); err != nil {
-				c.logger.WithError(err).Error("could not flush queue")
-			}
-
-			if err := c.esClient.Close(); err != nil {
-				c.logger.WithError(err).Error("could not close client connection")
-			}
-			c.esClient.Stop()
-		}()
-
-		for doc := range c.pipeline.outputCh {
-
-			c.esClient.Add(c.indexName, config.tzpe, doc)
-
-			select {
-			case <-ctx.Done():
-				c.logger.WithError(ctx.Err()).Error("closing log pipeline")
-				return ctx.Err()
-			default:
-			}
-		}
-		return nil
-	})
-
-	return nil
-}
-
 // StopLogging implements the docker plugin interface
 func (d *Driver) StopLogging(file string) error {
 
@@ -455,9 +141,4 @@ func (d *Driver) StopLogging(file string) error {
 	// }
 
 	return nil
-}
-
-// Name return the docker plugin name
-func (d Driver) Name() string {
-	return name
 }
