+++ pkg/docker/log_message.go
@@ -1,0 +1,74 @@
+package docker
+
+import (
+	"encoding/json"
+	"time"
+
+	"github.com/docker/docker/api/types/plugins/logdriver"
+	"github.com/docker/docker/daemon/logger"
+)
+
+// LogMessage ...
+type LogMessage struct {
+	logdriver.LogEntry
+	logger.Info
+
+	GrokLine map[string]string
+}
+
+// MarshalJSON ...
+func (l LogMessage) MarshalJSON() ([]byte, error) {
+	return json.Marshal(
+		struct {
+
+			// docker/daemon/logger/Info
+			Config              map[string]string `json:"config,omitempty"`
+			ContainerID         string            `json:"containerID,omitempty"`
+			ContainerName       string            `json:"containerName,omitempty"`
+			ContainerEntrypoint string            `json:"containerEntrypoint,omitempty"`
+			ContainerArgs       []string          `json:"containerArgs,omitempty"`
+			ContainerImageID    string            `json:"containerImageID,omitempty"`
+			ContainerImageName  string            `json:"containerImageName,omitempty"`
+			ContainerCreated    *time.Time        `json:"containerCreated,omitempty"`
+			ContainerEnv        []string          `json:"containerEnv,omitempty"`
+			ContainerLabels     map[string]string `json:"containerLabels,omitempty"`
+			LogPath             string            `json:"logPath,omitempty"`
+			DaemonName          string            `json:"daemonName,omitempty"`
+
+			//  api/types/plugin/logdriver/LogEntry
+			Line     string    `json:"message,omitempty"` // []byte to string
+			Source   string    `json:"source"`
+			TimeNano time.Time `json:"timestamp"` // int64 to Time
+			Partial  bool      `json:"partial"`
+
+			GrokLine map[string]string `json:"grok,omitempty"`
+		}{
+			Config:              l.Config,
+			ContainerID:         l.ContainerID,
+			ContainerName:       l.ContainerName,
+			ContainerEntrypoint: l.ContainerEntrypoint,
+			ContainerArgs:       l.ContainerArgs,
+			ContainerImageID:    l.ContainerImageID,
+			ContainerImageName:  l.ContainerImageName,
+			ContainerCreated:    l.timeOmityEmpty(),
+			ContainerEnv:        l.ContainerEnv,
+			ContainerLabels:     l.ContainerLabels,
+			LogPath:             l.LogPath,
+			DaemonName:          l.DaemonName,
+
+			GrokLine: l.GrokLine,
+
+			Line:     string(l.Line),
+			Source:   l.Source,
+			TimeNano: time.Unix(0, l.TimeNano),
+			Partial:  l.Partial,
+		})
+
+}
+
+func (l LogMessage) timeOmityEmpty() *time.Time {
+	if l.ContainerCreated.IsZero() {
+		return nil
+	}
+	return &l.ContainerCreated
+}
